Removed the need of a preference activity as it is now merged into the existing preferences
