+++ app/src/main/AndroidManifest.xml
@@ -358,9 +358,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".Modules.CEPASReader.activity.PreferencesActivity"
-            android:label="@string/preferences" />
-        <activity
             android:name=".Modules.CEPASReader.activity.SupportedCardsActivity"
             android:label="@string/supported_cards"
             android:screenOrientation="sensorPortrait" />
