--- app/src/main/res/menu/module_cepas_main.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-  main_menu.xml
-
-  Copyright (C) 2011 Eric Butler
-
-  Authors:
-  Eric Butler <eric@codebutler.com>
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-  -->
-
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/prefs"
-        android:title="@string/preferences"
-        android:icon="@android:drawable/ic_menu_preferences" />
-</menu>
